// src/types/PlayerStats.ts
interface PlayerStats {
    games_played: number;
    sum_minutes: number;
    sum_points: number;
    sum_goals: number;
    sum_assists: number;
    sum_bps: number;
    sum_cleansheets: number;
    sum_defensive_contributions: number;
    sum_xg: number;
    sum_xa: number;
    sum_xgi: number;
    sum_shots_on_target: number;
    sum_big_chances_created: number;
    sum_key_passes: number;
    sum_xgap: number;
}

export default PlayerStats;